pyproject.toml
[tool.poetry]
name = "algorithm"
version = "0.2.0"
description = ""
authors = ["diklios <47837462+diklios5768@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

# 清华
[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "default"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"


[[tool.poetry.source]]
name = "aliyun-source"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "supplemental"

[[tool.poetry.source]]
name = "pytorch-cu117"
url = "https://download.pytorch.org/whl/cu117"
priority = "explicit"

[[tool.poetry.source]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
# tools
python-dotenv = "^1.0.0"
click = "^8.1.3"
tqdm = "^4.64.0"
typer = { extras = ["all"], version = "^0.7.0" }
joblib = "^1.2.0"
munch = "^2.5.0"
numpy = "1.24.0"
# poetry plugin
poethepoet = "^0.16.2"
# network
requests = "^2.28.1"
# table
tables = "^3.8.0"
pandas = "^1.5.3"
openpyxl = "^3.0.10"
xlsxwriter = "^3.0.3"
prettytable = "^3.4.1"
# jupyter
jupyterlab = "^3.4.5"
jupyter = "^1.0.0"
notebook = "^6.4.12"
# fig
matplotlib = "3.5.3"
seaborn = "^0.11.2"
visdom = "^0.1.8.9"
pydotplus = "^2.0.2"
graphviz = "^0.20.1"
bokeh = "^3.3.0"
# pdf
PyPDF2 = "^2.10.4"
# image
opencv-python = "4.5.5.64"
Pillow = "^9.5.0"
imutils = "^0.5.4"
# nlp
transformers = "^4.20.1"
jieba = "^0.42.1"
simpletransformers = "^0.64.3"
# ml
scikit-learn = "^1.3.0"
scikit-image = "^0.19.3"
keras = "^2.9.0"
scipy = "^1.8.1"
sympy = "^1.10.1"
fastai = "^2.7.11"
mlxtend = "^0.21.0"
scanpy = "^1.9.3"
# dl
einops = "^0.4.1"
wandb = "^0.13.3"
tensorboard = "^2.10.0"
tensorboardX = "^2.5.1"
tensorlayer = "^2.2.5"
# torch
albumentations = "^1.3.0"
accelerate = "^0.20.3"
timm = "^0.9.2"
torch = "1.13.1"
torchgeometry = "^0.1.2"
torchsummary = "^1.5.1"
torchkeras = "^3.2.3"
torchsampler = "^0.1.2"
pytorch-lightning = "^1.9.4"
segmentation-models-pytorch = "^0.3.3"
# Interpretability
grad-cam = "^1.4.6"
torchcam = "^0.3.2"
captum = "^0.6.0"
shap = "^0.42.1"
lime = "^0.2.0.1"
# openmmlab
openmim = "^0.3.1"
# gpu
gpustat = "^1.0.0"

[tool.poe.tasks.torch-cpu]
cmd = "pip3 install torch torchvision torchaudio"

[tool.poe.tasks.torch-cpu-conda]
cmd = "conda install pytorch torchvision torchaudio cpuonly -c pytorch"

[tool.poe.tasks.torch-gpu]
#在使用这个命令之前先去网站查看一下torch版本对应有的cuda版本
cmd = "pip install torch==${torch_version} torchvision torchaudio --force --extra-index-url https://download.pytorch.org/whl/${CUDA}"
args = [
    { name = "torch_version", default = "2.0.0", positional = false, help = "torch version" },
    { name = "CUDA", default = "cu116", positional = true, help = "CUDA version" }
]

[tool.poe.tasks.torch-gpu-conda]
cmd = "conda install pytorch torchvision torchaudio cudatoolkit=${CUDA} -c pytorch -c conda-forge"
args = [
    { name = "CUDA", default = "11.6", positional = true, help = "CUDA version" }
]

[tool.poe.tasks.mmcv]
cmd = "mim install mmcv-full"

[tool.poe.tasks.tensorflow]
cmd = "pip install --upgrade tensorflow -i https://pypi.tuna.tsinghua.edu.cn/simple"

[tool.poe.tasks.R]
cmd = "conda install r-irkernel -c r -y"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
